lambda-srcs/sagemaker-deployer-function/requirements.txt
sagemaker
PyGithub
urllib3==1.26.6 
requests==2.27.1